[1.5 -> 1.8] Navigation Link & ScrollView / Profile View / First Refactor and Project Organization
NavigationLink between our LocationListView and LocationDetailView. We also throw the checked in users grid into a ScrollView.

Build the top section of the profile screen. In part two we will build the rest of the screen

Finish up the initial UI for our profile screen

First round of refactoring to clean up a bit. We also start organizing our files into folders